Add a debug helper for reading traffic with Wireshark
Wireshark can already helpfully dissect the initial SSH handshake. When given the
session keys, it can also dissect the encrypted traffic for inspection/debugging.
This adds a helper in Debug mode to write out that information in the format
Wireshark requires.

Usage is to set `Renci.SshNet.Abstractions.DiagnosticAbstraction.SOMETHING TODO`
and supply the same value to Wireshark in Edit -> Preferences -> Protocols -> SSH

The format was read out from
https://gitlab.com/wireshark/wireshark/-/blob/10d70661343a2548b5867bfd5a1964def327e938/epan/dissectors/packet-ssh.c#L2038